fks.lib の使用を止め、旧fname.cをfks.libと関数名を互換にしfks_fname.cppとして使用するように変更. vc10,vc14でのビルドを一応用意.

+++ src/app/jpgfileToPdf.cpp
@@ -134,7 +134,7 @@ bool JpgFileToPdf::addJpgPage(HPDF_Doc pdf, char const* jpgname, HPDF_Outline ro
                 HPDF_Page_SetFontAndSize(page, font, 9);
         }
         char str[130];  // 文字列化したint整数分には十分なサイズを指定.
-        sprintf(str, "%d", page_count_);
+        sprintf(str, "%d", (int)page_count_);
         HPDF_Outline     outline = HPDF_CreateOutline (pdf, root, str, NULL);
         HPDF_Destination dst = HPDF_Page_CreateDestination (page);
         HPDF_Destination_SetXYZ(dst, 0, HPDF_Page_GetHeight(page), 1);
